Added Network Class and associated changes

+++ src/simulation/Location.java
@@ -4,11 +4,11 @@ package simulation;
  * Tuple usable as neuron location
  */
 public class Location {
-    public int x = 0;
-    public int y = 0;
-    public int z = 0;
+    public double x = 0;
+    public double y = 0;
+    public double z = 0;
 
-    public Location(int x, int y, int z){
+    public Location(double x, double y, double z){
         this.x = x;
         this.y = y;
         this.z = z;
